requirements.txt
appdirs==1.4.3
asgiref==3.2.9
attrs==19.1.0
beautifulsoup4==4.9.1
black==19.3b0
boto3==1.14.43
botocore==1.17.43
certifi==2019.3.9
chardet==3.0.4
Click==7.0
Django==3.0.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.2.0
django-crispy-forms==1.9.1
django-storages==1.9.1
docutils==0.15.2
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.7.0
jmespath==0.10.0
Pillow==6.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.15.0
soupsieve==1.9.1
sqlparse==0.3.0
toml==0.10.0
urllib3==1.25.3
zipp==3.1.0